+++ story-service-impl/src/main/java/com/duan/story/dao/BaseDao.java
@@ -21,7 +21,7 @@ public interface BaseDao<T> {
      * @param id id
      * @return 操作影响的行数
      */
-    int delete(Long id);
+    int delete(Integer id);
 
     /**
      * 新增数据
