+++ story-service-impl/src/main/java/com/duan/story/dao/ProfileDao.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface ProfileDao extends BaseDao<Profile> {
 
-    Profile findProfileByWriterId(Long writerId);
+    Profile findProfileByWriterId(Integer writerId);
 
     Profile findProfileByPhone(String phone);
 }
