+++ story-service-impl/src/main/java/com/duan/story/dao/StoryStatisticsDao.java
@@ -13,17 +13,17 @@ import java.util.List;
 @Repository
 public interface StoryStatisticsDao extends BaseDao<StoryStatistics> {
 
-    StoryStatistics findStoryStatisticsByStoryId(Long storyId);
+    StoryStatistics findStoryStatisticsByStoryId(Integer storyId);
 
-    int updateCommentCount_plus(Long storyId);
-    int updateViewCount_plus(Long storyId);
-    int updateCollectCount_plus(Long storyId);
-    int updateLikeCount_plus(Long storyId);
-    int updateCommentCount_minus(Long storyId);
-    int updateCollectCount_minus(Long storyId);
-    int updateLikeCount_minus(Long storyId);
+    int updateCommentCount_plus(Integer storyId);
+    int updateViewCount_plus(Integer storyId);
+    int updateCollectCount_plus(Integer storyId);
+    int updateLikeCount_plus(Integer storyId);
+    int updateCommentCount_minus(Integer storyId);
+    int updateCollectCount_minus(Integer storyId);
+    int updateLikeCount_minus(Integer storyId);
 
-    Integer findViewCountByStoryId(Long storyId);
-    Integer findLikeCountByStoryId(Long storyId);
+    Integer findViewCountByStoryId(Integer storyId);
+    Integer findLikeCountByStoryId(Integer storyId);
 
 }
