+++ story-service-impl/src/main/java/com/duan/story/entity/Label.java
@@ -13,9 +13,9 @@ import java.sql.Timestamp;
 @Data
 public class Label implements Serializable {
 
-    private static final long serialVersionUID = 4565919455090875775L;
-    private Long id;
-    private Long writerId;
+    private static final Long serialVersionUID = 4565919455090875775L;
+    private Integer id;
+    private Integer writerId;
     private String title;
     private Timestamp createDate;
