+++ story-service/src/main/java/com/duan/story/common/dto/LinkDTO.java
@@ -12,9 +12,9 @@ import java.io.Serializable;
 @Data
 public class LinkDTO implements Serializable {
 
-    private static final long serialVersionUID = 5545136415441492665L;
-    private Long id;
-    private Long writerId;
+    private static final Long serialVersionUID = 5545136415441492665L;
+    private Integer id;
+    private Integer writerId;
     private String icon;
     private String title;
     private String url;
