+++ story-service/src/main/java/com/duan/story/common/dto/ProfileDTO.java
@@ -12,9 +12,9 @@ import java.io.Serializable;
 @Data
 public class ProfileDTO implements Serializable {
 
-    private static final long serialVersionUID = 6674849351249573711L;
-    private Long id;
-    private Long writerId;
+    private static final Long serialVersionUID = 6674849351249573711L;
+    private Integer id;
+    private Integer writerId;
     private String avatar;
     private String phone;
     private String email;
